+++ src/testing/test-helper.js
@@ -419,7 +419,7 @@ function verifyAccessAssignments(expectedAccessAssignments) {
   }
 }
 
-function verifyOperationChannelsAssigned(doc, oldDoc, expectedChannels) {
+function verifyOperationChannelsAssigned(doc, expectedChannels) {
   if (exports.channel.callCount !== 1) {
     assert.fail('Document channels were not assigned');
   }
@@ -474,7 +474,7 @@ function verifyAuthorization(expectedAuthorization) {
 }
 
 function verifyDocumentAccepted(doc, oldDoc, expectedAuthorization, expectedAccessAssignments) {
-  exports.syncFunction(doc, oldDoc);
+  exports.syncFunction(doc, oldDoc || null);
 
   if (expectedAccessAssignments) {
     verifyAccessAssignments(expectedAccessAssignments);
@@ -482,11 +482,11 @@ function verifyDocumentAccepted(doc, oldDoc, expectedAuthorization, expectedAcce
 
   const expectedOperationChannels = verifyAuthorization(expectedAuthorization);
 
-  verifyOperationChannelsAssigned(doc, oldDoc, expectedOperationChannels);
+  verifyOperationChannelsAssigned(doc, expectedOperationChannels);
 }
 
 function verifyDocumentCreated(doc, expectedAuthorization, expectedAccessAssignments) {
-  verifyDocumentAccepted(doc, void 0, expectedAuthorization || defaultWriteChannel, expectedAccessAssignments);
+  verifyDocumentAccepted(doc, null, expectedAuthorization || defaultWriteChannel, expectedAccessAssignments);
 }
 
 function verifyDocumentReplaced(doc, oldDoc, expectedAuthorization, expectedAccessAssignments) {
@@ -500,7 +500,7 @@ function verifyDocumentDeleted(oldDoc, expectedAuthorization, expectedAccessAssi
 function verifyDocumentRejected(doc, oldDoc, docType, expectedErrorMessages, expectedAuthorization) {
   let syncFuncError = null;
   try {
-    exports.syncFunction(doc, oldDoc);
+    exports.syncFunction(doc, oldDoc || null);
   } catch (ex) {
     syncFuncError = ex;
   }
@@ -516,7 +516,7 @@ function verifyDocumentRejected(doc, oldDoc, docType, expectedErrorMessages, exp
 }
 
 function verifyDocumentNotCreated(doc, docType, expectedErrorMessages, expectedAuthorization) {
-  verifyDocumentRejected(doc, void 0, docType, expectedErrorMessages, expectedAuthorization || defaultWriteChannel);
+  verifyDocumentRejected(doc, null, docType, expectedErrorMessages, expectedAuthorization || defaultWriteChannel);
 }
 
 function verifyDocumentNotReplaced(doc, oldDoc, docType, expectedErrorMessages, expectedAuthorization) {
@@ -593,7 +593,7 @@ function verifyAccessDenied(doc, oldDoc, expectedAuthorization) {
 
   let syncFuncError = null;
   try {
-    exports.syncFunction(doc, oldDoc);
+    exports.syncFunction(doc, oldDoc || null);
   } catch (ex) {
     syncFuncError = ex;
   }
@@ -637,7 +637,7 @@ function verifyAccessDenied(doc, oldDoc, expectedAuthorization) {
 function verifyUnknownDocumentType(doc, oldDoc) {
   let syncFuncError = null;
   try {
-    exports.syncFunction(doc, oldDoc);
+    exports.syncFunction(doc, oldDoc || null);
   } catch (ex) {
     syncFuncError = ex;
   }