+++ test/access-assignment.spec.js
@@ -44,10 +44,10 @@ describe('User and role access assignment:', function() {
       testHelper.verifyDocumentReplaced(doc, oldDoc, 'write', expectedStaticAssignments);
     });
 
-    it('is applied when deleting an existing document', function() {
+    it('is NOT applied when deleting an existing document', () => {
       var oldDoc = { _id: 'staticAccessDoc' };
 
-      testHelper.verifyDocumentDeleted(oldDoc, 'write', expectedStaticAssignments);
+      testHelper.verifyDocumentDeleted(oldDoc, 'write');
     });
 
     it('is NOT applied when creating an invalid document', function() {
@@ -128,34 +128,40 @@ describe('User and role access assignment:', function() {
       testHelper.verifyDocumentReplaced(doc, oldDoc, 'write', expectedDynamicAssignments);
     });
 
-    it('is applied when deleting an existing document', function() {
-      var oldDoc = { _id: 'dynamicAccessDoc' };
+    it('is applied even when the roles and users are not defined', () => {
+      const doc = { _id: 'dynamicAccessDoc' };
 
-      var expectedDeleteAssignments = [
+      const expectedTestAssignments = [
         {
           expectedType: 'channel',
           expectedUsers: null,
-          expectedChannels: [ doc._id + '-channel3' ]
-        },
-        {
-          expectedType: 'channel',
-          expectedRoles: null,
-          expectedChannels: [ doc._id + '-channel4' ]
-        },
-        {
-          expectedType: 'channel',
-          expectedUsers: null,
-          expectedRoles: null,
-          expectedChannels: [ doc._id + '-channel2', doc._id + '-channel1' ]
+          expectedChannels: [ `${doc._id}-channel3` ]
         },
         {
           expectedType: 'role',
           expectedUsers: null,
           expectedRoles: null
+        },
+        {
+          expectedType: 'channel',
+          expectedRoles: null,
+          expectedChannels: [ `${doc._id}-channel4` ]
+        },
+        {
+          expectedType: 'channel',
+          expectedUsers: null,
+          expectedRoles: null,
+          expectedChannels: [ `${doc._id}-channel2`, `${doc._id}-channel1` ]
         }
       ];
 
-      testHelper.verifyDocumentDeleted(oldDoc, 'write', expectedDeleteAssignments);
+      testHelper.verifyDocumentCreated(doc, 'write', expectedTestAssignments);
+    });
+
+    it('is NOT applied when deleting an existing document', () => {
+      const oldDoc = { _id: 'dynamicAccessDoc' };
+
+      testHelper.verifyDocumentDeleted(oldDoc, 'write');
     });
 
     it('is NOT applied when creating an invalid document', function() {